refactor: merge test files
